+++ src/ops.rs
@@ -18,6 +18,7 @@ use std::path::PathBuf;
 use is_executable::is_executable;
 use std::collections::BTreeMap;
 use std::fs::{self, File};
+use git2::Repository;
 
 pub fn do_build(env: &AmbosoEnv, args: &Args) -> Result<String,String> {
     match args.tag {
@@ -813,6 +814,51 @@ pub fn run_test(test_path: &PathBuf, record: bool) -> Result<String,String> {
 }
 
 pub fn gen_c_header(target_path: &PathBuf, target_tag: &String, bin_name: &String) -> Result<String,String> {
+    let repo = Repository::discover(target_path);
+    let mut head_author_name = "".to_string();
+    let id;
+    let commit_time;
+    match repo {
+        Ok(r) => {
+            let head = r.head();
+            match head {
+                Ok(head) => {
+                    let commit = head.peel_to_commit();
+                    match commit {
+                       Ok(commit) => {
+                           id = commit.id().to_string();
+                           info!("Commit id: {{{}}}", id);
+                           let author = commit.author();
+                           let name = author.name();
+                           match name {
+                              Some(name) => {
+                                   head_author_name = name.to_string();
+                                   info!("Commit author: {{{}}}", head_author_name);
+                               }
+                               None => {
+                                   warn!("Commit author is empty: {}", head_author_name);
+                               }
+                            }
+                            commit_time = commit.time().seconds();
+                            info!("Commit time: {{{}}}", commit_time);
+                               }
+                               Err(e) => {
+                                   error!("Failed peel to head commit for {{{}}}. Err: {e}", target_path.display());
+                                   return Err("Failed peel to head commit for repo".to_string());
+                               }
+                            }
+                }
+                Err(e) => {
+                    error!("Failed getting head for {{{}}}. Err: {e}", target_path.display());
+                    return Err("Failed getting head for repo".to_string());
+                }
+            }
+        }
+        Err(e) => {
+            error!("Failed discovering repo for {{{}}}. Err: {e}", target_path.display());
+            return Err("Failed discover of repo".to_string());
+        }
+    }
     let header_path = format!("{}/anvil__{}.h", target_path.display(), bin_name);
     trace!("Generating C header. Target path: {{{}}} Tag: {{{}}}", header_path, target_tag);
     let output = File::create(header_path);
@@ -822,9 +868,9 @@ pub fn gen_c_header(target_path: &PathBuf, target_tag: &String, bin_name: &Strin
 #define ANVIL__{bin_name}__\n
 static const char ANVIL__API_LEVEL__STRING[] = \"1.9.6\"; /**< Represents amboso version used for [anvil__{bin_name}.h] generated header.*/\n
 static const char ANVIL__{bin_name}__VERSION_STRING[] = \"{target_tag}\"; /**< Represents current version for [anvil__{bin_name}.h] generated header.*/\n
-static const char ANVIL__{bin_name}__VERSION_DESC[] = \"\"; /**< Represents current version info for [anvil__{bin_name}.h] generated header.*/\n
-static const char ANVIL__{bin_name}__VERSION_DATE[] = \"\"; /**< Represents date for current version for [anvil__{bin_name}.h] generated header.*/\n
-static const char ANVIL__{bin_name}__VERSION_AUTHOR[] = \"\"; /**< Represents author for current version for [anvil__{bin_name}.h] generated header.*/\n
+static const char ANVIL__{bin_name}__VERSION_DESC[] = \"{id}\"; /**< Represents current version info for [anvil__{bin_name}.h] generated header.*/\n
+static const char ANVIL__{bin_name}__VERSION_DATE[] = \"{commit_time}\"; /**< Represents date for current version for [anvil__{bin_name}.h] generated header.*/\n
+static const char ANVIL__{bin_name}__VERSION_AUTHOR[] = \"{head_author_name}\"; /**< Represents author for current version for [anvil__{bin_name}.h] generated header.*/\n
 const char *get_ANVIL__API__LEVEL__(void); /**< Returns a version string for amboso API of [anvil__{bin_name}.h] generated header.*/\n
 const char *get_ANVIL__VERSION__(void); /**< Returns a version string for [anvil__{bin_name}.h] generated header.*/\n
 const char *get_ANVIL__VERSION__DESC__(void); /**< Returns a version info string for [anvil__{bin_name}.h] generated header.*/\n